Lecture13/requirements.txt
arize-phoenix-otel
arize-phoenix-client
python-dotenv
openai
agno
rich